requirements.txt
transformers
pytorch-crf
sacremoses
torch
mlflow
peft

